refactor: 카탈로그 아이템 변환 유틸 정리
